+++ src/DbEx.Postgres/Resources/ExtendedSchema/Stored Procedures/sp_throw_authorization_exception.sql
@@ -1,0 +1,15 @@
+-- Copyright (c) Avanade. Licensed under the MIT License. See https://github.com/Avanade/DbEx
+
+CREATE OR REPLACE PROCEDURE sp_throw_authorization_exception(
+  "message" TEXT = NULL
+)
+LANGUAGE plpgsql
+AS $$
+BEGIN
+  IF "message" IS NULL THEN
+    "message" := '';
+  END IF;
+
+  RAISE USING MESSAGE = "message", ERRCODE = '56003';
+END
+$$;