+++ src/DbEx.SqlServer/Resources/ExtendedSchema/Stored Procedures/spThrowAuthorizationException.sql
@@ -1,0 +1,9 @@
+-- Copyright (c) Avanade. Licensed under the MIT License. See https://github.com/Avanade/DbEx
+
+CREATE OR ALTER PROCEDURE [dbo].[spThrowAuthorizationException]
+  @Message NVARCHAR(2048) = NULL
+AS
+BEGIN
+  SET NOCOUNT ON;
+  THROW 56003, @Message, 1
+END